+++ app/src/main/java/woowacourse/shopping/base/BaseActivity.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.base
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.common.AppViewModelFactory
+import woowacourse.shopping.di.AppViewModelFactory
 
 abstract class BaseActivity : AppCompatActivity() {
     override val defaultViewModelProviderFactory: ViewModelProvider.Factory by lazy { AppViewModelFactory() }
